[#46] 🩹 fix : 문제 댓글 구조 변경
